+++ src/main/java/com/weatherbydegys/backend/repos/FavCityRepo.java
@@ -1,13 +1,12 @@
 package com.weatherbydegys.backend.repos;
 
 import com.weatherbydegys.backend.model.FavCity;
-import com.weatherbydegys.backend.model.User;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
 public interface FavCityRepo extends CrudRepository<FavCity, Integer>, JpaSpecificationExecutor<FavCity> {
-    FavCity findById(int id);
+    FavCity findById(long id);
     List<FavCity> findAll();
 }
